Refactor output rendering code
