+++ Nodexr/wwwroot/css/site.css
@@ -257,11 +257,11 @@ input {
     margin: 5px;
 }
 
-    .output-regex-container:first-child {
+    .output-regex-container :first-child {
         border-radius: 5px 0 0 5px;
     }
 
-    .output-regex-container:last-child {
+    .output-regex-container :last-child {
         border-radius: 0 5px 5px 0;
     }
